requirements.txt
python-telegram-bot==20.7
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
schedule==1.2.1
feedparser==6.0.11
praw==7.7.1
yfinance==0.2.36
textblob==0.18.0
google-generativeai==0.8.3

